Move notifications handling to a separate module
